refactor: separate relay failure diagnostics
