+++ supabase/functions/relay/requestGate.ts
@@ -32,18 +32,6 @@ export type RelayRequestSlot =
 
 export const RELAY_SLOT_REFRESH_INTERVAL_MS = 60_000;
 
-export type RelayFailurePhase =
-  'pre_headers_timeout' | 'pre_headers_failure' | 'response_body_failure';
-
-const UPSTREAM_REQUEST_ID_HEADERS = [
-  'request-id',
-  'x-request-id',
-  'x-goog-request-id',
-] as const;
-
-type RelayRequestBody =
-  string | Uint8Array | ReadableStream<Uint8Array> | null | undefined;
-
 class RelayRequestSlotLostError extends Error {
   constructor() {
     super('relay_request_refresh did not find request slot');
@@ -70,38 +58,6 @@ function parseGateDecision(data: unknown): GateDecision {
 
 function didRefreshSlot(data: unknown): boolean {
   return isJsonRecord(data) && data.refreshed === true;
-}
-
-export function classifyPreHeaderFailure(error: unknown): RelayFailurePhase {
-  return error instanceof Error &&
-    (error.name === 'TimeoutError' || error.name === 'AbortError')
-    ? 'pre_headers_timeout'
-    : 'pre_headers_failure';
-}
-
-export function getUpstreamRequestId(headers: Headers): string | null {
-  for (const header of UPSTREAM_REQUEST_ID_HEADERS) {
-    const requestId = headers.get(header)?.trim();
-    if (requestId) return requestId;
-  }
-  return null;
-}
-
-export function getRelayRequestBytes(
-  body: RelayRequestBody,
-  contentLength: string | null,
-): number | null {
-  if (body == null) return 0;
-  if (typeof body === 'string') {
-    return new TextEncoder().encode(body).byteLength;
-  }
-  if (body instanceof Uint8Array) return body.byteLength;
-
-  if (contentLength == null || contentLength.trim() === '') return null;
-  const parsedLength = Number(contentLength);
-  return Number.isSafeInteger(parsedLength) && parsedLength >= 0
-    ? parsedLength
-    : null;
 }
 
 function once(release: () => Promise<void>): () => Promise<void> {
@@ -200,7 +156,7 @@ export async function releaseWhenStreamCloses(
   release: () => Promise<void>,
   refresh?: () => Promise<void>,
   refreshIntervalMs = RELAY_SLOT_REFRESH_INTERVAL_MS,
-  onUpstreamBodyError?: (error: unknown) => void,
+  onUpstreamBodyError?: () => void,
 ): Promise<ReadableStream<Uint8Array> | null> {
   if (body === null) {
     await release();
@@ -245,7 +201,7 @@ export async function releaseWhenStreamCloses(
         readResult = await reader.read();
       } catch (error) {
         try {
-          onUpstreamBodyError?.(error);
+          onUpstreamBodyError?.();
         } catch {
           console.error('[RELAY] Upstream body failure observer failed');
         }
